requirements.txt
fastapi
langchain
langchain-community
sentence-transformers
huggingface-hub
faiss-cpu
python-dotenv
pydantic